Moved Document structs to a separate file for better organization.
